pyproject.toml
[project]
name = "ai-sports-outcome-predictor"
version = "1.0.0"
description = "Lightweight, LLM-first Telegram sports prediction bot (no DB, no heavy ML)"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.7",
    "aiohttp>=3.11.0",
    "requests>=2.32.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.5.0",
    "google-generativeai>=0.8.3",
    "mcp[cli]>=0.9.1",
    "browser-use>=0.1.0",
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "langchain-openai>=0.3.31",
    "langchain-google-genai>=2.0.0",
    "anthropic==0.40.0",
    "pydantic>=2.0.0",
    "duckduckgo-search>=6.1.0",
    "beautifulsoup4>=4.12.0",
    "ddgs>=1.0.0"
]
